Troubleshoot accuracy readings ...

+++ manager/aica_django/microagents/AICAFlowerClient.py
@@ -149,7 +149,7 @@ class AICAFlowerClient(NumPyClient):  # type: ignore
                 print(y_train.shape)
                 optimizer.zero_grad()
                 outputs = self.aica_model(train_data)
-                loss = criterion(outputs, train_labels.type(torch.LongTensor))
+                loss = criterion(outputs, train_labels.type(torch.FloatTensor))
                 loss.backward()
                 optimizer.step()
 
@@ -186,7 +186,7 @@ class AICAFlowerClient(NumPyClient):  # type: ignore
                 emb = X_batch.to(self.device)
                 labels = y_batch.to(self.device)
                 outputs = self.aica_model(emb)
-                loss += criterion(outputs, labels.type(torch.LongTensor)).item()
+                loss += criterion(outputs, labels.type(torch.FloatTensor)).item()
                 correct += (
                     (torch.max(outputs.data, 1)[1] == torch.max(labels, 1)[1])
                     .sum()
